+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Support for TM1638 seven segment display by Ajith Vasudevan (#11031)
 - Support for MAX7219 seven segment display by Ajith Vasudevan (#11387)
 - Support for Frequency monitoring and zero-cross detection on CSE7761 (Sonoff Dual R3)
+- IRremoteESP8266 library from v2.7.15 to v2.7.16
 
 ### Changed
 - PubSubClient library from EspEasy v2.7.12 to Tasmota v2.8.12